Add files via upload
In this notebook, we explore a sequence defined recursively by $a_{n+2}=0.5(a_n+a_{n+1})$. We argue for its convergence in the reals and compute its limit in a few ways (some mentioned in passing while others given in the appendices). We also explore its convergence and see the results via a plot as well as the listing down of the first few terms, the number of terms required for the desired accuracy is computed exactly - and displayed.