Add files via upload
This note studies the few famous inequalities surrounding distances or l_p-norms. l_p-norms are indispensable tools in machine learning. We do not study the use of l_p-norms in machine learning in this note, rather we study its properties - learning about the dial to adjust for data good fit.